Added single tests to "ant test"
